+++ android/app/build.gradle
@@ -47,7 +47,7 @@ android {
 }
 
 dependencies {
-    compile "com.facebook.react:react-native:0.20.1"
+    compile "com.facebook.react:react-native:+"
     compile 'com.nimbusds:nimbus-jose-jwt:5.3'
     compile "com.google.code.gson:gson:2.8.2"
     compile 'com.microsoft.appcenter:appcenter:1.2.0'
