--- android/app/src/test/java/com/microsoft/codepush/common/TestUtils.java
@@ -1,28 +1,0 @@
-package com.microsoft.codepush.common;
-
-/**
- * Utils to make testing process easier and avoid code repetition.
- */
-class TestUtils {
-    final static String CLIENT_UNIQUE_ID = "YHFv65";
-    final static String DEPLOYMENT_KEY = "ABC123";
-    final static String LABEL = "awesome package";
-    final static boolean FAILED_INSTALL = false;
-    final static boolean IS_PENDING = true;
-    final static boolean IS_DEBUG_ONLY = false;
-    final static boolean IS_FIRST_RUN = false;
-    final static String PREVIOUS_DEPLOYMENT_KEY = "prevABC123";
-    final static String PREVIOUS_LABEL = "awesome package previous";
-    final static String APP_VERSION = "2.2.1";
-    final static String STATUS = "Succeeded";
-    final static String DESCRIPTION = "short description";
-    final static boolean IS_MANDATORY = true;
-    final static boolean UPDATE_APP_VERSION = true;
-    final static boolean IS_AVAILABLE = true;
-    final static boolean SHOULD_RUN_BINARY = false;
-    final static long PACKAGE_SIZE = 102546723;
-    final static String DOWNLOAD_URL = "https://url.com";
-    final static String PACKAGE_HASH = "HASH";
-    final static String ERROR = "An error has occurred";
-    final static String SERVER_URL = "https";
-}
